# Arbcall structs