--- a/src/precision.jl
+++ b/src/precision.jl
@@ -1,10 +1,16 @@
 const DEFAULT_PRECISION = Ref{Clong}(256)
 
 """
-    precision(<:Union{Arf,Arb,Acb})
+    precision(<:Union{Arf, Arb, Acb, arf_struct, arb_struct, acb_struct})
+    precision(<:Ptr{<:Union{arf_struct, arb_struct, acb_struct}})
+    precision(x::Union{arf_struct, arb_struct, acb_struct})
+    precision(x::Ptr{<:Union{arf_struct, arb_struct, acb_struct}})
 Get the default precision (in bits) currently used for `Arblib` arithmetic.
 """
-Base.precision(::Type{<:Union{Arf,Arb,Acb}}) = DEFAULT_PRECISION[]
+Base.precision(::Type{<:Union{Arf, Arb, Acb, arf_struct, arb_struct, acb_struct}}) = DEFAULT_PRECISION[]
+Base.precision(::Type{<:Ptr{<:Union{arf_struct, arb_struct, acb_struct}}}) = DEFAULT_PRECISION[]
+Base.precision(x::Union{arf_struct, arb_struct, acb_struct}) = DEFAULT_PRECISION[]
+Base.precision(x::Ptr{<:Union{arf_struct, arb_struct, acb_struct}}) = DEFAULT_PRECISION[]
 
 """
     setprecision(::Type{<:Union{Arf, Arb, Acb}}, precision::Int)