--- a/src/precision.jl
+++ b/src/precision.jl
@@ -9,8 +9,11 @@ Get the default precision (in bits) currently used for `Arblib` arithmetic.
 """
 Base.precision(::Type{<:Union{Arf, Arb, Acb, arf_struct, arb_struct, acb_struct}}) = DEFAULT_PRECISION[]
 Base.precision(::Type{<:Ptr{<:Union{arf_struct, arb_struct, acb_struct}}}) = DEFAULT_PRECISION[]
+
 Base.precision(x::Union{arf_struct, arb_struct, acb_struct}) = DEFAULT_PRECISION[]
 Base.precision(x::Ptr{<:Union{arf_struct, arb_struct, acb_struct}}) = DEFAULT_PRECISION[]
+
+Base.precision(x::Union{Arf, Arb, Acb}) = x.prec
 
 """
     setprecision(::Type{<:Union{Arf, Arb, Acb}}, precision::Int)
@@ -26,4 +29,10 @@ function Base.setprecision(::Type{<:Union{Arf,Arb,Acb}}, precision::Integer)
     end
     DEFAULT_PRECISION[] = precision
     return precision
+end
+
+function Base.setprecision(x::T,
+                           precision::Integer;
+                           shallow = false) where {T <: Union{Arf, Arb, Acb}}
+    return T(x, prec = precision, shallow = shallow)
 end