--- a/src/types.jl
+++ b/src/types.jl
@@ -73,11 +73,6 @@ mutable struct Acb <: Number
     end
 end
 
-cprefix(::Type{Mag}) = :mag
-cprefix(::Type{Arf}) = :arf
-cprefix(::Type{Arb}) = :arb
-cprefix(::Type{Acb}) = :acb
-
 for (T, prefix) in ((Mag, :mag), (Arf, :arf), (Arb, :arb), (Acb, :acb))
     arbstruct = Symbol(prefix, :_struct)
     spref = "$prefix"
@@ -85,6 +80,7 @@ for (T, prefix) in ((Mag, :mag), (Arf, :arf), (Arb, :arb), (Acb, :acb))
         cstructtype(::Type{$T}) = $arbstruct
     end
     @eval begin
+        cprefix(::Type{$T}) = $(QuoteNode(Symbol(prefix)))
         cstruct(x::$T) = getfield(x, cprefix($T))
         Base.convert(::Type{$(cstructtype(T))}, x::$T) = cstruct(x)
     end