--- a/test/arb_types-test.jl
+++ b/test/arb_types-test.jl
@@ -1,16 +1,8 @@
 @testset "arb_types" begin
-    mag = Arblib.mag_struct(0,0)
-    arf = Arblib.arf_struct(0,0,0,0)
-    arb = Arblib.arb_struct(0,0,0,0,0,0)
-    acb = Arblib.acb_struct(0,0,0,0,0,0, 0,0,0,0,0,0)
-
-    for x in (mag, arf, arb, acb)
-        @test begin
-            Arblib.init!(x)
-            Arblib.clear!(x)
-            true
-        end
-    end
+    mag = Arblib.mag_struct()
+    arf = Arblib.arf_struct()
+    arb = Arblib.arb_struct()
+    acb = Arblib.acb_struct()
 
     prec = 256
     for x in (arf, arb, acb)