--- a/test/runtests.jl
+++ b/test/runtests.jl
@@ -3,6 +3,7 @@ using Arblib, Test
 include("arb_types-test.jl")
 include("types-test.jl")
 include("arbcall-test.jl")
+include("precision-test.jl")
 include("constructors-test.jl")
 include("predicates-test.jl")
 include("show-test.jl")