I have changed the methods generated by `arbcall` to not only accept the Julia wrapper of the Arb-type but also the corresponding struct as well as pointers to such structs. So for example an argument which would previously take an `Arb` now takes `Union{Arb, arb_struct, Ptr{arb_struct}}`. If the argument is a struct or a pointer the default precision is the current global precision.

The main reason for this is to make it easier to implement non-allocating and inplace operations on vectors and matrices, but also on the radius and midpoint, in the future.

Some more comments:

The last commit fixes a performance issue with `cprefix` that I noticed when comparing the performance after the two commits before that. I believe the problem was that the symbol had to be parsed every time. I tried to figure out how to precompute it in the macro but was not able to get anywhere... So I moved it out and defined it manually instead.

There are not many tests for the new code, in particular for calling the methods with pointers. This will probably be added in a later stage when we add support for `ArbVectors` which will naturally have to use a lot of the pointer methods.