@kalmarek 2020-05-30
just from the philosophical point of view: I do agree with the idea of passing "naked" Arb structs (or pointers to them) to the functions we define. 
But maybe a better idea is to channel everything through `Ptr{arb_struct}` and (as @saschatimme suggested) attach finalizers there and make `Arb` immutable?

This would require overloading `cconvert(::Type{Ref{arb_struct}}, a::Arb) = cconvert(::Type{Ref{arb_struct}}, a.arb)` but that's easy to do :)

eg.
```julia
struct Arb <: Real
    arb::arb_struct
    prec::Int

    Arb(;prec::Integer=DEFAULT_PRECISION[]) = new(arb_struct(), prec)
end

mutable struct arb_struct
                        # ┌ arf_struct (midpoint)
    exponent::UInt      # │ fmpz
    size::UInt          # │ mp_size_t
    mantissa1::UInt     # │ mantissa_struct of length 128
    mantissa2::UInt     # │
                        # └
                        # ┌ mag_struct (radius)
    exponent_mag::UInt  # │ fmpz
    mantissa_mag::UInt  # │ mp_limb_t
                        # └

    function arb_struct()
        a = new()
        init!(a)
        finalizer(clear!, a)
        return a
    end
end


ctype(::Carg{T}) where T <: Union{Arb}  = Ref{arb_struct}
Base.cconvert(::Type{Ref{arb_struct}}, x::Arb) = Base.cconvert(Ref{arb_struct}, cstruct(x))
```

and

```julia
julia> using Arblib, BenchmarkTools

julia> a = Arblib.Arb("0.1")
[0.1000000000000000000000000000000000000000000000000000000000000000000000000000 +/- 1.95e-78]

julia> b = zero(a)
0

julia> Arblib.add!(b, a,a); b
[0.20000000000000000000000000000000000000000000000000000000000000000000000000000 +/- 3.89e-78]

julia> @benchmark Arblib.add!($b,$a,$a)
BenchmarkTools.Trial: 
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     51.657 ns (0.00% GC)
  median time:      53.995 ns (0.00% GC)
  mean time:        57.738 ns (0.00% GC)
  maximum time:     291.661 ns (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     985

```

julia-1.3.1:
```julia
julia> @benchmark Arblib.add!($b,$a,$a)
BenchmarkTools.Trial: 
  memory estimate:  48 bytes
  allocs estimate:  3
  --------------
  minimum time:     250.930 ns (0.00% GC)
  median time:      267.055 ns (0.00% GC)
  mean time:        289.759 ns (0.24% GC)
  maximum time:     2.768 μs (83.74% GC)
  --------------
  samples:          10000
  evals/sample:     301

```

julia-1.5.0-beta1.0:
```julia
julia> @benchmark Arblib.add!($b,$a,$a)
BenchmarkTools.Trial: 
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     51.575 ns (0.00% GC)
  median time:      53.408 ns (0.00% GC)
  mean time:        60.592 ns (0.00% GC)
  maximum time:     233.736 ns (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     985
```