@kalmarek 2020-05-30
> Julia 1.5 :heart_eyes: 

it's not different from 1.4.2 in this task :stuck_out_tongue: