@saschatimme 2020-05-30
So, I think the best strategy would be then this (just for the `arb` type for simplicity):

1) `arb_struct` is a `mutable struct` and has the finalizer
2) `Arb` is `struct` holding `arb_struct` and `precision::Int`
3) Our functions work with `Union{Arb,arb_struct,Ptr{arb_struct}}`.

Only critical part is how to handle the precision if the assigned value is not an `Arb`. Maybe we require it to be always an `Arb`? Since in 1.5 these can guaranteed be created without any allocations from an `arb_struct` this seems reasonable in my opinion.

Thoughts?