@kalmarek 2020-05-30
another takes on precision:
* if the first argument is an `Arb`, then it has it's own precision (and we take it as default); in other cases we require precision to be explicitly specified, or
* define `precision(::arb_struct) = 0` and take `max` of precision of args, throwing an error if it is too small (which forces the user to explicitly pass the precision in the case only `arb_structs are passed`

currently I'm not a big fan of auto-precision, except maybe for our constructors, but that's not a breaking point for me either.