@Joel-Dahne 2020-05-31
I like your ideas! Regarding the precision I would be fine with having to explicitly give the precision for `arb_struct` and `Ptr{arb_struct}`, I see them as a low level interface and then it's better to be explicit with them. Since the precision at the moment is only determined by the mutated argument we can easily dispatch on that for handling the different versions.