@Joel-Dahne 2020-05-31
Some comments from trying to implement this:

Making `Arb` immutable is problematic because you can no longer change the precision. It's quite common with Arb that you iterate a computation, doubling the precision every time until you get good enough results.

With the code
```julia
using Arblib, BenchmarkTools
a, b = Arb(), Arb();
@benchmark Arblib.add!($b, $a, $a)
```
I get the following results with Julia 1.3.1 (1.4.2 seems to be the same)

On master:
```
BenchmarkTools.Trial: 
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     11.563 ns (0.00% GC)
  median time:      11.574 ns (0.00% GC)
  mean time:        11.761 ns (0.00% GC)
  maximum time:     96.975 ns (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     999
```
With the discussed changes (not pushed here yet):
```
BenchmarkTools.Trial: 
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     11.564 ns (0.00% GC)
  median time:      11.574 ns (0.00% GC)
  mean time:        11.681 ns (0.00% GC)
  maximum time:     34.066 ns (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     999

```
With the above discussed changes but also making `Arb` mutable:
```
BenchmarkTools.Trial: 
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     13.111 ns (0.00% GC)
  median time:      13.129 ns (0.00% GC)
  mean time:        13.438 ns (0.00% GC)
  maximum time:     45.376 ns (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     998
```
The timings seems to be a bit unstable but making `Arb` mutable does seem to make things slightly slower.