@saschatimme 2020-05-31
> Making Arb immutable is problematic because you can no longer change the precision. It's quite common with Arb that you iterate a computation, doubling the precision every time until you get good enough results.

I think that’s not much of a problem since you can just construct a new `Arb` from the old one with different precision. The underlying arb memory is still the same.