@saschatimme 2020-05-31
Also the `2.5ns` are in line with the cost of the additional pointer indirection.