@Joel-Dahne 2020-05-31
At the moment we do not have any non-copying constructor for `Arb`, so you cannot create a new `Arb` with updated precision without having to allocate first. We can of course implement it, but I'm not sure how we want to do it, we probably want
```julia
x = Arb()
y = Arb(x, prec = 64)
```
to copy `x`.