@saschatimme 2020-05-31
Yes, that's true. But I think we could have a `set_precision(x::Arb, prec::Int)` function which calls an (internal) non-copying constructor.