@kalmarek 2020-05-31
ah yes, but since we change into more `arb_struct` oriented design we should probably add new constructors: `Arb(as::arb_struct, prec::Integer)` and maybe even `Arb(a::Arb, prec::Integer)` for setting precision.

I'm surprised that your julia 1.3 doesn't allocate though @Joel-Dahne ?
btw. is my hardware so slow? :D how fast is yours macbook @saschatimme? (5x faster is impressive).